Removed warning.

+++ cuda/UT/UTDecoding/src/UTDecodeRawBanksInOrder.cu
@@ -89,7 +89,7 @@ __global__ void ut_decode_raw_banks_in_order(
 
     // const uint32_t highThreshold = threshold;
     const uint32_t channelStripID = chanID + strip;
-    const uint32_t LHCbID = (0xB << 28) | channelStripID;
+    const uint32_t LHCbID = (((uint32_t) 0xB) << 28) | channelStripID;
     // const uint32_t planeCode = 2 * station + (layer & 1);
 
     ut_hits.yBegin[hit_index] = yBegin;
